chore/doc: update documentation and fill in missing gaps
